requirements.txt
telethon
pillow
python-dotenv
aiosqlite
pyahocorasick
imagehash
piexif
pytz